short-circuit actions and middlewares if headers are sent
With the current official version of routing-controllers v0.7.7, there is an issue
 where two routes could match the same path, ie:

/resource/non-dynamic-value
/resource/:id

and both methods and their respective sets of middlewares will run. This
leads to race conditions, and express throwing errors of the type:
headers cannot be set after being sent.

In this commit we remove this issue by short-circuiting actions and
middlewares if the headers have already been sent.

The one downside to this approach is that afterMiddlewares will no longer run.